refactor: Authenticator 테스트 app -> data 모듈로 이동
